+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed 
 
 ### Added 
-	
+- Added TROPOMI Level 2 reader	
 ### Changed 
 
 ### Removed