+++ CHANGELOG.md
@@ -5,9 +5,33 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [Unreleased] - yyyy-mm-dd 
+# [Unreleased] - yyyy-mm-dd
+
+### Fixed
+### Added
+### Changed
+
+### Removed
+
+### Deprecated
+
+# [v1.2.2] - 2025-03-03 
 
 ### Fixed 
+- bad reshape import in kde.py
+### Added 
+- Added TROPOMI Level 2 reader	
+### Changed 
+
+### Removed 
+
+### Deprecated 
+
+
+# [v1.2.1] - 2025-02-04 
+
+### Fixed 
+- Bug fix in aop.py for returning PMOM
 
 ### Added